# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test ApplicationAPI",
        "description": "Documentação Swagger no Laravel",
        "contact": {
            "email": "rangelthr@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autenticado",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Deslogado com sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Informe o token"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/transaction": {
            "get": {
                "tags": [
                    "transaction"
                ],
                "description": "Retorna todas as transações do usuários",
                "operationId": "App\\Http\\Controllers\\TransactionController::index",
                "responses": {
                    "200": {
                        "description": "Transações retornados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "transaction"
                ],
                "description": "Criando nova transação",
                "operationId": "App\\Http\\Controllers\\TransactionController::store",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Valor da transação",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "document",
                        "in": "query",
                        "description": "Documento do beneficiário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registrado com sucesso."
                    },
                    "422": {
                        "description": "Dados ausentes ou inconsistentes."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user_category": {
            "get": {
                "tags": [
                    "user category"
                ],
                "description": "Retorna todas as categorias de usuário",
                "operationId": "App\\Http\\Controllers\\UserCategoryController::index",
                "responses": {
                    "200": {
                        "description": "Transações retornados com sucesso."
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "user"
                ],
                "description": "Retorna todos os usuários",
                "operationId": "App\\Http\\Controllers\\UserController::index",
                "responses": {
                    "200": {
                        "description": "Usuários retornados com sucesso."
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user"
                ],
                "description": "Criando novo usuário",
                "operationId": "App\\Http\\Controllers\\UserController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nome completo do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id_user_category",
                        "in": "query",
                        "description": "Categorias ( ID: 1 - Comum ou 2 - Lojistas)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "document",
                        "in": "query",
                        "description": "Documento",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Senha",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Registrado com sucesso."
                    },
                    "422": {
                        "description": "Dados ausentes."
                    }
                }
            }
        },
        "/api/user/{user}": {
            "get": {
                "tags": [
                    "user"
                ],
                "description": "Retorna o usuário com id especificado",
                "operationId": "App\\Http\\Controllers\\UserController::show",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário retornado com sucesso."
                    },
                    "422": {
                        "description": "Usuário informado não encontrado."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/wallet/{user}": {
            "get": {
                "tags": [
                    "wallet"
                ],
                "description": "Retorna a carteira do usuário",
                "operationId": "App\\Http\\Controllers\\WalletController::show",
                "responses": {
                    "200": {
                        "description": "Carteira do usuário informada, retorna informações"
                    },
                    "422": {
                        "description": "Usuário informado não encontrado."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}